+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/k8-automation/compare/0.5.2...HEAD
 
+### Changed
+
+-   Attempt to always include error information on failure of build
+    handler
+
+### Fixed
+
+-   Support GitHub.com owners and repositories that do not conform to
+    docker image name and tag restrictions
+
 ## [0.5.2][] - 2018-03-03
 
 [0.5.2]: https://github.com/atomist/k8-automation/compare/0.5.1...0.5.2